INT-4616 - Stage and Application drop downs do not work with Jenkins plugin 3.11.202103 and newer - enhance tests of the pipeline syntax page
